requirements.txt
requests
dash
dash-core-components
dash-html-components
plotly
pandas
spotipy
python-dotenv
